+++ tudanumafes2019/tudanumafes2019/Assets/Scenes/DifferentScene.unity
@@ -336,6 +336,7 @@ GameObject:
   - component: {fileID: 149826844}
   - component: {fileID: 149826846}
   - component: {fileID: 149826845}
+  - component: {fileID: 149826847}
   m_Layer: 5
   m_Name: Canvas
   m_TagString: Untagged
@@ -459,6 +460,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: ad45724b1c75e0c4da86ce27bd56d969, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &149826847
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 149826839}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 86f7e6789b1526248827f5524a2aacc7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &251479385
