+++ tudanumafes2019/tudanumafes2019/Assets/Scenes/RuleScene.unity
@@ -283,6 +283,7 @@ GameObject:
   - component: {fileID: 1408174858}
   - component: {fileID: 1408174857}
   - component: {fileID: 1408174861}
+  - component: {fileID: 1408174862}
   m_Layer: 5
   m_Name: Canvas
   m_TagString: Untagged
@@ -380,7 +381,19 @@ MonoBehaviour:
   m_GameObject: {fileID: 1408174856}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ad45724b1c75e0c4da86ce27bd56d969, type: 3}
+  m_Script: {fileID: 11500000, guid: 4be9e8f3ae6aa69488ae5f83a9ee9436, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &1408174862
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1408174856}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4be9e8f3ae6aa69488ae5f83a9ee9436, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &1631538626
