+++ tudanumafes2019/tudanumafes2019/Assets/Scenes/Script/ChangeScript.cs
@@ -1,0 +1,37 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+using UnityEngine.SceneManagement;
+
+public class ChangeScript : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (Input.GetKey(KeyCode.Escape)) Quit();
+
+        if (Input.GetKey(KeyCode.Space)) Change();
+
+    }
+
+    public void Quit()
+    {
+#if UNITY_EDITOR
+        UnityEditor.EditorApplication.isPlaying=false;
+#else
+        Application.Quit();
+#endif
+    }
+
+    public void Change()
+    {
+        SceneManager.LoadScene("TitleScene");
+    }
+}
