+++ tudanumafes2019/tudanumafes2019/Assets/Scenes/Script/ChangeScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86f7e6789b1526248827f5524a2aacc7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
