--- tudanumafes2019/tudanumafes2019/Assets/Scenes/Script/EffectScript.cs
@@ -1,22 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class EffectScript : MonoBehaviour
-{
-    public int rotspeed = 20;
-
-
-    // Start is called before the first frame update
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-
-    }
-}
