+++ tudanumafes2019/tudanumafes2019/Assets/Scenes/Script/RuleScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 96bc365022ee9f14ca0fae2101ad1283
+guid: 4be9e8f3ae6aa69488ae5f83a9ee9436
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
