+++ tudanumafes2019/tudanumafes2019/Assets/Scenes/TitleScene.unity
@@ -906,6 +906,18 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ad45724b1c75e0c4da86ce27bd56d969, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &2024124676
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2024124667}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 86f7e6789b1526248827f5524a2aacc7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &2024124677
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1050,7 +1062,7 @@ PrefabInstance:
     - target: {fileID: 317150798435281703, guid: b6a46bbcb44f07043aaa9cf30db074d3,
         type: 3}
       propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.size
-      value: 1
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 317150798435281703, guid: b6a46bbcb44f07043aaa9cf30db074d3,
         type: 3}
@@ -1126,7 +1138,17 @@ PrefabInstance:
         type: 3}
       propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[1].m_Target
       value: 
-      objectReference: {fileID: 2024124667}
+      objectReference: {fileID: 2024124676}
+    - target: {fileID: 317150798435281703, guid: b6a46bbcb44f07043aaa9cf30db074d3,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[1].m_MethodName
+      value: Quit
+      objectReference: {fileID: 0}
+    - target: {fileID: 317150798435281703, guid: b6a46bbcb44f07043aaa9cf30db074d3,
+        type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[1].m_Arguments.m_ObjectArgumentAssemblyTypeName
+      value: UnityEngine.Object, UnityEngine
+      objectReference: {fileID: 0}
     - target: {fileID: 317150799872264216, guid: b6a46bbcb44f07043aaa9cf30db074d3,
         type: 3}
       propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.size
